+++ src/utility/log/Log.cc
@@ -39,7 +39,7 @@ namespace ieda {
  * @param data
  * @param size
  */
-void SignalHandle(const char* data, int size)
+void SignalHandle(const char* data, size_t size)
 {
   std::ofstream fs("glog_dump.log", std::ios::app);
   std::string str = std::string(data, size);
